Done with homework2

+++ homework2/cookie-clicker-lab4-7/src/main/java/edu/csula/web/AdminEventsRemoveServlet.java
@@ -11,9 +11,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import edu.csula.storage.UsersDAO;
 import edu.csula.storage.servlet.EventsDAOImpl;
 import edu.csula.storage.EventsDAO;
 import edu.csula.models.Event;
+import edu.csula.storage.servlet.UsersDAOImpl;
 
 @WebServlet("/admin/events/remove")
 
@@ -22,9 +24,15 @@ public class AdminEventsRemoveServlet extends HttpServlet {
     @Override
     public void doGet( HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         // TODO: handle upsert transaction
-        EventsDAO dao = new EventsDAOImpl(getServletContext());
-        int id=Integer.parseInt(request.getParameter("id"));
-        dao.remove(id);
-        response.sendRedirect("/admin/events");
+        UsersDAO dao1=new UsersDAOImpl(request.getSession());
+        if(dao1 != null) {
+            EventsDAO dao = new EventsDAOImpl(getServletContext());
+            int id = Integer.parseInt(request.getParameter("id"));
+            dao.remove(id);
+            response.sendRedirect("/admin/events");
+        }
+        else{
+            dao1.logout();
+        }
     }
 }
